[#2] Added basic player movement
The OrbitalShield moves clockwise if the player presses the right side of the screen, and counterclockwise if they press on the left side of the screen.

The LeftRightInputManager script exposes an event that is invoked when the pressed side changes. The Orbital script implements an interface and is hooked into the manager's event through the Editor.

Closes #2